+++ src/oldnews/commands/ui.py
@@ -6,6 +6,13 @@ from textual_enhanced.commands import Command
 
 
 ##############################################################################
+class Escape(Command):
+    """Back out through the panes, or exit the app if the navigation pane has focus"""
+
+    BINDING_KEY = "escape, q"
+
+
+##############################################################################
 class ToggleShowAll(Command):
     """Toggle between showing all and showing only unread"""
 
@@ -17,6 +24,27 @@ class ToggleCompact(Command):
     """Toggle a more compact user interface"""
 
     BINDING_KEY = "f5"
+
+
+##############################################################################
+class JumpToSubscriptions(Command):
+    """Jump to the subscriptions panel"""
+
+    BINDING_KEY = "1"
+
+
+##############################################################################
+class JumpToArticles(Command):
+    """Jump to the articles panel"""
+
+    BINDING_KEY = "2"
+
+
+##############################################################################
+class JumpToArticle(Command):
+    """Jump to the article panel"""
+
+    BINDING_KEY = "3"
 
 
 ### ui.py ends here